fix:modelの修正
