+++ src/main.go
@@ -1,9 +1,4 @@
 package main
 
-import (
-	"github.com/NenfuAT/xr-project-xrStudyWatch-back/lib"
-)
-
 func main() {
-	lib.SqlConnect()
 }
