+++ src/model/table.go
@@ -1,0 +1,5 @@
+package model
+
+import "github.com/NenfuAT/xr-project-xrStudyWatch-back/lib"
+
+var db = lib.SqlConnect()
