--- src/models/location.go
@@ -1,7 +1,0 @@
-package model
-
-type Location struct {
-	ID       string `gorm:"primarykey;type:varchar(26)" json:"id"`
-	Building string `json:"building"`
-	Room     string `json:"room"`
-}
