--- src/models/university.go
@@ -1,7 +1,0 @@
-package model
-
-type University struct {
-	ID           string `json:"id"`
-	UniversityID string `json:"universityId"`
-	Name         string `json:"name"`
-}
